+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - 2026-09-22
+
+### Added
+- Made resolved text links in language-server hover previews clickable.
+
+### Changed
+- Rendered filesystem links in language-server hover previews as inline code.
+
 ## [0.12.0] - 2026-09-22
 
 ### Added